Add First-Pass Mouse Controls (See #5)
Includes basic click and drag functionality.  Some bugs include slightly inaccurate raycasting/dragging, particularly when the cursor is to the left of a splitter or pusher.  Another major bug is when two splitters/pushers are dragged on top of each other, becoming a single splitter/pusher since there was never any collision detection between the two.